feat: [대시보드] Backend 로직 설계
- ClubRepository.java : long -> int 변경
- BoardCategoryRepository.java : count 큰 순서 추가
- DashboardService.java : 타입 변경 및 로직 추가

+++ crewz-admin/src/main/java/crewz/admin/crewzadmin/repository/UserRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import crewz.admin.crewzadmin.model.entity.User;
 
 public interface UserRepository extends JpaRepository<User, Long> {
-	long countBy();
+	int countBy();
 }
